refactor sample opportunities route out of global config/state land
