+++ packages/openactive-broker-microservice/src/sample-opportunities.js
@@ -1,18 +1,28 @@
 const { getRelevantOffers } = require('@openactive/test-interface-criteria');
-const { HARVEST_START_TIME } = require('./broker-config');
 
 // NOTE: duplicated from openactive-integration-tests/test/helpers/flow-stages/fetch-opportunities.js
-function getRandomRelevantOffer(opportunity, criteriaName) {
+/**
+ * @param {string} harvestStartTime
+ * @param {import('./models/core').Opportunity} opportunity
+ * @param {string} criteriaName
+ */
+function getRandomRelevantOffer(harvestStartTime, opportunity, criteriaName) {
   const relevantOffers = getRelevantOffers(criteriaName, opportunity, {
-    harvestStartTime: HARVEST_START_TIME,
+    harvestStartTime,
   });
   if (relevantOffers.length === 0) { return null; }
 
   return relevantOffers[Math.floor(Math.random() * relevantOffers.length)];
 }
 
-function renderSampleOpportunities(opportunity, criteriaName, sellerId) {
-  const offer = getRandomRelevantOffer(opportunity, criteriaName);
+/**
+ * @param {Pick<import('./broker-config').BrokerConfig, 'HARVEST_START_TIME'>} config
+ * @param {import('./models/core').Opportunity} opportunity
+ * @param {string} criteriaName
+ * @param {string} sellerId
+ */
+function renderSampleOpportunities(config, opportunity, criteriaName, sellerId) {
+  const offer = getRandomRelevantOffer(config.HARVEST_START_TIME, opportunity, criteriaName);
 
   return {
     sampleOpportunities: [
